requirements.txt
flask==2.0.3
aioflask
pyjwt==2.3.0
requests==2.24.0
nlpcloud~=1.0.23
itsdangerous==2.0.1
gunicorn
python-dotenv
machaao==0.3.5
configparser~=5.2.0
nltk
transformers==4.29.2
torch
accelerate
peft
sentencepiece